major update to species.db (not done) and other additions

+++ PokeBattleTest/date.h
@@ -1796,7 +1796,7 @@ low_level_fmt(std::basic_ostream<CharT, Traits>& os, const year& y)
     detail::save_ostream<CharT, Traits> _(os);
     os.fill('0');
     os.flags(std::ios::dec | std::ios::internal);
-    os.width(4 + (y < year{0}));
+    os.width(4 + static_cast<int>((y < year{0})));
     os.imbue(std::locale::classic());
     os << static_cast<int>(y);
     return os;
